Added projects app
Added/configured projects app
