+++ website_project/django_project/urls.py
@@ -28,8 +28,7 @@ urlpatterns = [
     path('thanks/', views.ThanksPage.as_view(), name='thanks'),
     path('posts/', include('posts.urls', namespace='posts')),
     path('groups/', include('groups.urls', namespace='groups')),
-    path('projects/', views.ProjectPage.as_view(), name='projects'),
-    path('projects/extra-life/', views.ExtraLife.as_view(), name='extralife'),
+    path('projects/', include("projects.urls")),
 ]
 
 if settings.DEBUG:
